hook: add sign-off using "interpret-trailers"
The sample hook to prepare the commit message before a commit allows
users to opt-in to add the sign-off to the commit message.  The
sign-off is added at a place that isn't consistent with the "-s"
option of "git commit".  Further, it could go out of view in certain
cases.

Add the sign-off in a way similar to "-s" option of "git commit"
using git's interpret-trailers command.

"interpret-trailers" alone works well in all cases except when the
user invokes "git commit" without any arguments.  In that case
manually add a new line after the first line to ensure it's
consistent with the output of "-s" option.

Signed-off-by: Kaartic Sivaraam <kaarticsivaraam91196@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>